Updated everything but task 6

+++ src/GeometricFigure.java
@@ -1,4 +1,4 @@
 public interface GeometricFigure {
 
-    public int calculateArea();
+    public double calculateArea();
 }
